+++ crates/state/src/session.rs
@@ -296,7 +296,10 @@ impl Session {
     pub fn submit_input(&mut self, text: String, cx: &mut Context<Self>) {
         if let Some(input) = &self.input {
             input.send(text).ok();
-            if let SessionState::Authorizing(Some(SignInPrompt::Secret)) = &self.state {
+            if let SessionState::Authorizing(Some(
+                SignInPrompt::Secret | SignInPrompt::Accounts(_),
+            )) = &self.state
+            {
                 self.state = SessionState::Authorizing(None);
                 cx.notify();
             }